tools/pyproject.toml
[tool.poetry]
name = "arxiv-keycloak-tools"
version = "0.1.0"
description = "Development tools for arXiv Keycloak"
authors = ["Naoyuki Tai <ntai@arxiv.org>"]
license = "BSD-3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "83af5ee568c5710ca3df0f6f68e39f998a772fa8"}
mysqlclient = "^2.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
